pyproject.toml
[tool.poetry]
name = "cygnss-wetlands"
version = "0.1.0"
description = "A repo dedicated to investigating global inundation patterns with NASA CYGNSS observations and other supporting data sources."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cygnss_wetlands"}]

[tool.poetry.dependencies]
python = "^3.9"
xarray = "^2023.3.0"
numpy = "^1.24.2"
rasterio = "^1.3.6"
geopandas = "^0.12.2"
pandas = "^1.5.3"
jupyterlab = "^3.6.2"
netCDF4 = "^1.6.3"
python-dotenv = "^1.0.0"
s3fs = "^2023.3.0"
tqdm = "^4.65.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
flake8 = "6.0.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
flake8-pyproject = "^1.2.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"

[tool.poetry.scripts]
cygnss = "cygnss_wetlands.cygnss.cli:entry"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = "\\.pyi?$"
exclude = "/(\n    \\.git\n  | \\.hg\n  | \\.mypy_cache\n  | \\.tox\n  | \\.venv\n  | _build\n  | buck-out\n  | build\n  | dist\n)/\n"

[tool.isort]
profile = "black"

[tool.flake8]
ignore = [ "E203", "E266", "E501", "F401", "F403", "W503",]
max-line-length = 120
count = true
select = [ "B", "C", "E", "F", "W", "T4", "B9",]
